Update README installation instructions and remove Python dependency
This commit updates the README installation instructions by moving the command to install thrift-compiler to the SDK development section. This is done to keep SDK specific instructions in one place and help facilitate SDK development. Furthermore, the Python dependency was removed from CMake to simplify the build process and reduce dependency on specific Python versions.