Adding `UpdateExpr` class and support for it in related operations
The addition introduces a new class, `UpdateExpr`, which encapsulates details of an update expression. Support for `UpdateExpr` was included in the related Thrift services and classes including the `ParsedExprType`, `UpdateRequest` and `RemoveThriftClient`. Methods to read and write these new fields are added as well. This change allowed for the handling of `between_expr` expressions in update operations. This is necessary to extend the functionality and allow more detailed control and variations of update operations in the RPC interactions.