Streamline error handling and update scripts in documentation
This commit not only simplifies the error handling in the file `sqllogictest.py` by eliminating redundant print statements but also increases the version of our project in the `pyproject.toml` file. Additionally, the instructions and command syntax have been updated in the benchmark documentation to use the new script names, specifying the optional parameter value for better clarity and accuracy.