Add block size constraint in insert test (#437)
Added an additional test case in test_insert.py to handle inserts that exceed the block size limit. Also, ensured that the table drop operation is successful in the test. Updated the physical_insert.cpp to throw an error when the number of insert rows is larger than the default block capacity.