[Fix bug] Deny the connect request after shutdown the server (#647)
### What problem does this PR solve?

When we use pg_client to connect infinity, due to the pg_clients haven't
disconnected with infinity. The infinity server can't be shutdown. At
this moment, infinity server can still be connected with pg_client.
Correct behavior should be infinity server should deny the access in
this moment, since infinity server is in stopping status.

Be aware, this PR is only fix the PG client connection. For thrift
server part, we haven't fixed it.

### Type of change

- [x] Bug Fix (non-breaking change which fixes an issue)

Signed-off-by: jinhai <haijin.chn@gmail.com>