Add flush option, but it is not used right now. (#795)
### What problem does this PR solve?

Currently, the log flush policy is fixed to Flush at once. This PR is to
add a config which can be used to change the flush policy. Be aware,
this PR only introduce the config, but setting the parameter won't
change the flush behavior right now.

In the future, another PR is need to activate this function.

### Type of change

- [x] New Feature (non-breaking change which adds functionality)

Signed-off-by: Jin Hai <haijin.chn@gmail.com>