Fix: conflict between cleanup and gc of buffer manager (#861)
### What problem does this PR solve?

Problem:
1. A `buffer_object` will appear multiple times in `gc_queue_` of
`BufferManager` , which is conflict with cleanup cause it will remove
`buffer_object` itself, thus `gc_queue_` will hold wild pointer.
2. Queue blocked: Currently, `wal_manager` will produce the checkpoint
task and send it to `background_processer` when the wal log size is too
big. This will cause the task queue in `background_processer` blocked.
Consider such a situation when the `background_processer` is handling a
checkpoint task, and the txn of checkpoint is waiting to commit. It need
`WalManager::Flush()` to commit txn, but in `Flush()` , a new checkpoint
task is submitting to `background_processer` but the queue is full. Thus
the queue blocked cause it need to commit the current txn to have free
slot but the txn need `Flush()` to handle itself which is waiting queue.
3. After cleaning up `table_entry` , the raw pointer saved in index file
worker will be wild pointers which will cause segment fault when
destructing itself.

Things have done:
1. Add a flag to ensure that each `buffer_object` appears only once in
the `gc_queue_` .
2. Add new status to express the status of ready for cleanup.
3. Add more unit test for new state machine, data insertion and so on.
4. Change the pointer to  `shared_ptr`  in index file worker.
5. Add a global variable in `WalManager` to judge whether a checkpoint
txn is being processed, ensuring that at most only one checkpoint
transaction is in progress.


Issue link:#849

### Type of change

- [x] Bug Fix (non-breaking change which fixes an issue)
- [x] Refactoring
- [x] Test cases

---------

Co-authored-by: shenyushi <shenyushi99@qq.com>