Fix: insufficient source data for MergeAggregate (#974)
### What problem does this PR solve?

The `MergeAggregate` operator may encounter errors due to receiving
insufficient data. This issue typically occurs when a single task within
a fragment is required to process multiple rounds of blocks. Currently,
when the `TaskScheduler` selects a task for execution, it only handles a
single block before invoking `PhysicalSink` , which generates a round of
`fragment_data`. Consequently, a task may produce more than one round of
`fragment_data` to send to `MergeAggregate` . However, in
`PhysicalSource` , it is assumed that one task corresponds to one round
of blocks.


Issue link:#962

### Type of change

- [x] Bug Fix (non-breaking change which fixes an issue)
- [x] Test cases