Fix the issue where items that failed during this clean could never be deleted (#2643)
### What problem does this PR solve?

Now, if the current clean has failed items, there will still be an opp
for those items to be deleted during next clean trigger. However, in
this fix, if the clean process contains failed items, then during the
next clean, all items (including those that were deleted) will be
retried. This could cause a perfermance issue when there is a large
clean metadata.

### Type of change

- [x] Bug Fix (non-breaking change which fixes an issue)
- [x] New Feature (non-breaking change which adds functionality)

---------

Signed-off-by: noob <yixiao121314@outlook.com>
Co-authored-by: Jin Hai <haijin.chn@gmail.com>