Bug Fix: buff obj state after save (#2882)
### What problem does this PR solve?

_Briefly describe what this PR aims to solve. Include background context
that will help reviewers understand the purpose of the PR._

Issue link:(https://github.com/infiniflow/infinity/issues/2871)

### Type of change

- [x] Bug Fix (non-breaking change which fixes an issue)

Before: during save, the buffer obj will change from kephermal to
kpersistent if it's saved whether it has later change or not. This will
cause an issue as the timestamp for txn and actual implementation time
are different. It's possible that an modification at a later timestamp
happens at an earlier time thus it won't be recorded in the file. And if
a file is deemed persistent and no modification is applied at a later
time, at next save, the data won't be saved again leading to data loss.

After: during save check if there are any modification after checkpoint
ts. If no we will make it kpersistent, or it will stay Kephermal. This
will make sure no real time and timestamp conflict affect persisting
data.

---------

Co-authored-by: joy-joy-joy-joy-joy-joy-joy <97589642+joy-joy-joy-joy-joy-joy-joy@users.noreply.github.com>