Fix meta cache issue (#2961)
### What problem does this PR solve?
**Issue**
The scenario with error is as below. Create index erases table cache,
and then show index adds a new table cache. Optimize index intends to
add index info to table cache, but it can not see index1, so the index
info in table cache is not latest. Then snapshot reads the wong index
info in table cache.

create  index (idx1)-------show index ------snapshot
  ** |---------------optimize-------------------------------|

**Solution**
   
We allow txn to change the table cache only when the txn begins after
all committed write txns.
If it is conflicted with any committed write txn, the info it intends to
change to the table cache may be not correct.

### Type of change

- [x] Bug Fix (non-breaking change which fixes an issue)