Fix "wal.log does not exist" issue (#3049)
### What problem does this PR solve?

delete_ut.cpp failed with error "wal.log is not found" intermittently. 

At the end of the testunit, db directory is removed explicitly. At the
same time wal manager is still working in NewFlush().
wal.log is needed in NewFlush().
We should not call RemoveDbDirs() at the end of the testunit, db
directory will be cleaned in TearDown().

### Type of change

- [x] Test cases