Refactor hnsw index dump in PopulateIndex (#3136)
### What problem does this PR solve?
When you create an index after appending data, the original
implementation calls PopulateIndexToMem, which launches a background
task that attempts to dump the in‑memory index five times. These
attempts always fail because the operation is still inside a
transaction, so the newly created index’s metadata is never visible.
Afterward, DumpSegmentIndex is invoked.

We can refactor the implementation to eliminate these unnecessary dump
attempts.

### Type of change

- [x] Refactoring
- [x] Performance Improvement

Signed-off-by: noob <yixiao121314@outlook.com>